+++ src/Stryker.Core/Stryker.Core.UnitTest/Mutants/TestResources/Mutator_SyntaxShouldBe_IfStatement_OUT.cs
@@ -9,12 +9,40 @@ namespace StrykerNet.UnitTest.Mutants.TestResources
         void TestMethod()
         {
             int i = 0;
-            if (System.Environment.GetEnvironmentVariable("ActiveMutation") == "0")
+            if (Stryker.ActiveMutationHelper.ActiveMutation==0)
             {
                 i = i - 1;
             } else {
                 i = i + 1;
             }
+            if (Stryker.ActiveMutationHelper.ActiveMutation==2)
+            {
+                string SomeLocalFunction()
+                {
+                    var test3 = 2 + 5;
+                    return $"test{1 - test3}";
+                }
+            }
+            else
+            {
+                if (Stryker.ActiveMutationHelper.ActiveMutation==1)
+                {
+                    string SomeLocalFunction()
+                    {
+                        var test3 = 2 - 5;
+                        return $"test{1 + test3}";
+                    }
+                }
+                else
+                {
+                    string SomeLocalFunction()
+                    {
+                        var test3 = 2 + 5;
+                        return $"test{1 + test3}";
+                    }
+                }
+            };
+            Console.WriteLine(SomeLocalFunction());
         }
     }
 }