+++ src/Stryker.Core/Stryker.Core/Mutants/MutantStatus.cs
@@ -6,7 +6,6 @@
         Killed,
         Survived,
         Timeout,
-        RuntimeError,
         BuildError
     }
 }